Update server.py
Thread - Safe Operation: The with lock: statement is used again to ensure that only one thread can access and modify the shared resources (tuple_space and total_ops) at a time.
READ Operation: If the command is R (representing a READ operation), the rest of the message is considered as the key. The code checks if the key exists in the tuple_space dictionary. If it does, the corresponding value is retrieved, and a success response message is constructed. The total_ops['READ'] counter is incremented to keep track of the number of successful READ operations. If the key does not exist, an error response message is constructed, and the total_ops['ERR'] counter is incremented.