Update server.py
This function is the entry point of the server program.
It prompts the user to enter a port number and creates a server socket.
It binds the server socket to the specified address (localhost) and port and starts listening for incoming connections.
It starts a daemon thread to run the log_stats function, which will periodically log server statistics.
It enters an infinite loop to accept incoming client connections. For each connection, it starts a new daemon thread to run the handle_client function.